refactor(db): move admin moderation store ownership
